Fixed flipper collisions and absorber issue

+++ src/model/LeftFlipperGizmo.java
@@ -22,13 +22,14 @@ public class LeftFlipperGizmo extends Gizmo{
 	}
 	
 	private void makeLeftFlipper(){
-		double radius = size/8;
+		double radius = (double) size/8;
 		int y2 = this.getEndY();
 		
 		corners.add(new Circle(x + (radius), y + (radius), radius));
 		corners.add(new Circle(x + (radius), y2 - (radius), radius));
 		lines.add(new LineSegment(x + (2 * radius), y + (radius), x + (2 * radius), y2 - (radius)));
 		lines.add(new LineSegment(x, y + (radius), x, y2 - (radius)));
+		
 	}
 	
 	public void rotateOnPivot(){
